Force focus on the main window as soon as it is opened
Note: This prevents the program from being closed with Alt+F4 before it gets the focus. If the program were closed this way, it would hang because the Alt+F4 event cannot be intercepted if the window is nor focused. By forcing focus on the window I can ensure that the program exits properly by using my clean_exit() function.